refactor: moved util code/ interface out of maphandler
