+++ Client/src/index.html
@@ -74,15 +74,15 @@
             <img src="" alt="signup image" class="signup-image" />
             <form class="signup-form">
               <div class="input-wrapper">
-                <label for="signupName">name</label>
+                <label for="signupName">이름</label>
                 <input id="signupName" class="signup-name" autocomplete="off" />
               </div>
               <div class="input-wrapper">
-                <label for="signupEmail">email</label>
+                <label for="signupEmail">이메일</label>
                 <input id="signupEmail" class="signup-email" autocomplete="off" />
               </div>
               <div class="input-wrapper">
-                <label for="signupPw">password</label>
+                <label for="signupPw">비밀번호</label>
                 <input
                   type="password"
                   id="signupPw"
@@ -91,7 +91,7 @@
                 />
               </div>
               <div class="input-wrapper">
-                <label for="signupPwConfirm">password confirm</label>
+                <label for="signupPwConfirm">비밀번호 확인</label>
                 <input
                   type="password"
                   id="signupPwConfirm"
@@ -114,12 +114,13 @@
           </button>
         </div>
         <div class="bookstore-info is-active">
+
           <div class="closest-store">
             <div class="loc-loader"></div>
             <span class="info">가까운 지점 찾기</span>
             <ol class="closest-store-list"></ol>
           </div>
-
+          <div class="spinner"></div>
           <div class="bookstore-map">
             <!-- Here's where 민기's magic happens -->
           </div>
@@ -139,6 +140,7 @@
             <span class="info">가까운 지점 찾기</span>
             <ol class="closest-store-list"></ol>
           </div>
+          <div class="spinner"></div>
           <div class="theatre-map">
             <!-- Here's where 민기's magic happens -->
           </div>
