+++ Client/src/scss/footer.scss
@@ -27,6 +27,7 @@
     }
     .indiego-info {
       width: 80vw;
+      max-width: 600px;
       text-align: justify;
       line-height: 2;
       font-size: 1.3rem;
