+++ Client/src/ts/interface/interface.ts
@@ -1,0 +1,15 @@
+export interface Branch {
+  id: number;
+  name: string;
+  address: string;
+  tel: string;
+  openhour: string;
+  introduction: string;
+  website: string;
+  instagram: string;
+  img: string;
+  lat: number;
+  lng: number;
+  distance: number;
+  region: string;
+}
